Fix Card component, rename it to Ticket
